Correct the API desc about the defaults for Full Ratios in OCS-Operator
While adding the fields the description of the fields were directly
lifted off from rook-operator. The values of NearFull, BackfillFull &
Full in rook are 0.85, 0.90 & 0.95 respectively. But in OCS Operator
we set these values to 0.75, 0.80 & 0.85 respectively with the help
of the rook-config-override ConfigMap. So the description of the fields
in the API should reflect the actual values that are set in OCS.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>