Include EiT state as part of the desired state hash sent to clients
When Encryption in-transit(EiT) is enabled/disabled the kernel mount
option for cephFS needs to be updated between prefer-crc/secure. So the
desired state hash needs to include the EiT state, so that
if the EiT state is changed the desired state hash will change and
the client will reconcile to get the updated mount option.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>