Defer secret hash update until all resources are successfully created
This update ensures that the secret hash in the storage
cluster status is updated only after all resources are created
successfully.
Previously, the status was updated prematurely, preventing the
reconcile loop from re-running whether there was a failure or success.
This change enables the reconcile process to continue retrying
until all resources are successfully created.

Signed-off-by: parth-gr <paarora@redhat.com>