Use TSCs & Always enforce them even if custom placement is specified
Currently TSCs are enforced only if there is no other placement spec
defined. Without the TSCs, the pods might not be distributed evenly.
So we should always enforce the TSCs even if some custom placement is
specified.

This also refactors the code to use TSCs in-place of Pod
affinity/anti-affinity for ceph daemons like mgr, mon, mds, rgw, nfs.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>