remove MDSCacheUsageHigh prometheus alert
Customers are complaining about erroneous MDS cache usage alerts.
Ceph team suggested that `ceph_mds_mem_rss` might not be the right
metric to capture this cache usage. So this alert needs to looked at
again. For the time being, we can just remove this alert due to
increasing number of customer cases around this.

Signed-off-by: Santosh <sapillai@redhat.com>