Remove ROOK_CSI_ENABLE_CEPHFS key from rook-ceph-operator-config CM
The ROOK_CSI_ENABLE_CEPHFS key was set in the external mode within the
rook-ceph-operator-config ConfigMap. However, it should have been
removed when transitioning from rook-ceph-operator-config to
ocs-operator-config. The rook-ceph-operator-config ConfigMap is now
retained only for customer modifications.

This code can be safely removed in the next release, as it only require
to run once to delete the key.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>