Use bulk flag for all odf data pools for performance gain
The bulk flag makes the autoscaler start with the max number of PGs,
and the autoscaler then decreases the pg count only if the PG usage
starts to skew too much. This could improve performance for ODF users
due to great amount of parallelism during read/write on large clusters.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>