Allow poolSpec for default block, file, object pools from storageCluster
With this change the poolSpec for the default cephBlockPool, default
cephFS data pool and default cephObjectStore data pool can be configured
via the storageCluster CR. The pool spec is directly passed to the ceph
CRs & only necessary fields are checked if set or not. Rest of the
fields are not touched. Although not all the fields are configurable,
but most fields are.

This will allow users to configure the target_size_ratio for these
pools via the storageCluster CR.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>