core: add back the CSI_DISABLE_HOLDER_PODS key on the cm
If the OCS operator is upgraded before the rook operator,
It will go and remove the CSI_DISABLE_HOLDER_PODS key
from the ocs-configmap,
Removing the key from the ocs-cm will re-start
the rook operator pods,

As Rook has not been upgraded and is still on 4.17,
it looks for the env variable
value which was coming from ocs-configmap key,
And as the key is deleted
rook operator got stuck.

Revert of the pr #2609

And Revert the fix in 4.19.

Signed-off-by: parth-gr <partharora1010@gmail.com>
(cherry picked from commit 843bd4c94cf3d46e6e76ad30731990ac44b4b12a)
Signed-off-by: parth-gr <partharora1010@gmail.com>