Fix conflict between SSD and HDD device classes causing CRUSH rule issue
In clusters where the devices were initially of type hdd and later
replaced with ssd, Ceph continues to list hdd in the available device
classes under ceph osd crush class ls. This occurs even after all
devices have been replaced with ssd. As rook populates the available
device classes from this list, the hdd device class persists in the
CephCluster status.

Starting from OCS 4.16, OCS operator determines and assigns a
defaultCephDeviceClass to all pools by selecting the first non-replica-1
entry in the CephCluster status. Due to the persistence of the hdd
device class, the OCS operator incorrectly selects hdd as the default
device class, even though all underlying device sets have been updated
to ssd.

To address this issue, we are introducing a hardcoded check in the OCS
operator: if both hdd and ssd device classes are present in the
CephCluster status, the operator will prioritize and select the ssd
device class as the default. Ideally Ceph should automatically
update the device class list when devices are replaced, so may be a fix
is required on ceph side as well.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>