nullable secret field in storageconsumer status and noobaaaccount fix
storageconsumer status field:
once the storageconsumer is enabled we shouldn't show secret field in
the status altogether, atm we just show the name w/ empty string and
this commit allows us to set the field to `nil`.

noobaaacount fix:
after consumer is enabled, client doesn't report status instantaneously
and if it's empty we reconcileNoobaaAccount once, which'll get into a
deadlock as noobaa will not be ready until we send storageconfig for
storageclass creation which will check for consumer readiness and it'll
never be ready.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>