Always apply rbd-mirror debug logging ceph configs
Irrespective of whether rbd-mirror is deployed or not, the debug logging
configs are always applied. This makes the code simpler considering we
have 2 ways of deploying rbd-mirror, one via the storageCluster
controller & another via the mirroring controller.
As without rbd-mirror present, the debug logging configs are anyways
not used so it is not a problem.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>