do not mark virt SC as default for kubevirt workloads
we were partially reconciling virt storageclass after it's creation and
this allowed external entities to remove the annotation which shouldn't
have been allowed.

with updated design ocs-operator holds full control over the
storageclass and this is disallowing user not to remove the annotation.
since we allow the addition of annotations we create the SC w/o the
annotation and user can mark it as default.

we are considering this as an interim fix to unblock CNV.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>