Set maintenance mode only when requested by at least one consumer
Previously we were setting maintenance mode based on the absence of
the rbd-mirror CR only, due to which maintenance mode was getting set
even in non rdr cluster. We now check if at least one consumer is
requesting maintenance mode based on the annotation & rbd-mirror cr
is not present before setting the maintenance mode.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>