Verify if the sub installedCSV is the expected one from package manifest
To solve the upgrade deadlock issue due to early sending of sub channel
we earlier introduced a check to verify if the sub installedCSV has
succeeded. But there is a brief window where the sub channel has been
updated but the sub installedCSV is not yet updated. In that window our
check will pass and we will send the new channel even though the upgrade
has not yet succeeded.

This commit adds a check to verify if the sub installedCSV is the
expected one by matching it with the csv name from the
package manifest's channel list. This helps us to avoid the earlier
race condition. We use a dynamic client to get the package manifest
as the typed client will not work for it.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>