Verify if the sub channel is valid with the package manifest
To solve the upgrade deadlock issue due to early sending of sub channel
we earlier introduced a check to verify if the sub installedCSV has
succeeded. But there is a brief window where the sub channel has been
updated but the sub installedCSV is not yet updated. In that window our
check will pass and we will send the new channel even though the upgrade
has not yet succeeded.

This commit adds a check to verify if the sub installedCSV is the
expected one by matching it with the csv name from the
package manifest's channel list. This helps us to avoid the earlier
race condition. We are using the unstructured object to get the
package manifest to avoid the added dependencies the import would bring.
Use mutex locks to lock & unlock the package level vars to avoid
concurrency issues with multiple rpc calls.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>