Filter packageManifests with the catalogSource Name & Namespace
We just discovered there can be multiple catalogSources with the same
name & namespace. So we need to filter the list with the catSrc Name &
Namespace of the subscription to get the correct packageManifest.
We will do the filtering on our side as we don't trust the backend
filtering as it already violates k8s standards.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>