Check volumeClaimTemplate & Resources spec are not empty before applying
From 4.18 to 4.19 There was a API change which converted the
sc.Spec.managedResources.cephNonResilientPools.VolumeClaimTemplate
& sc.Spec.ManagedResources.CephNonResilientPools.Resources fields
from a struct to a pointer to struct. Accordingly the code was modified.
But when someone upgrades from 4.18 to 4.19 the empty struct created
during 4.18 stays. Which negates the nil check being done on 4.19.
So operator believes there is some spec given to it & applies it while
in reality it's an empty struct. To fix this along with
a nil check we need to do a not empty check before applying the values.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>