Remove all other methods of installation apart from 'make install'
Having multiple methods of installation is leading to confusion and
inconsistency in the operator installation process. We had along with
the make install, a catalog based installation method, deploy-with-olm
yaml, and cluster-deploy tool as well. Having such many methods
is confusing and creates difficulties in maintaining the operator.
Hence, we are removing all the other methods of installation apart from
the 'make install' method. We will maintain this single method of
upstream installation going forward.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>