Specify the full ratios directly in the cephCluster CR
Specifying the full ratios in the cephConfig section is not supported.
It throws this error "failed to set config store options: failed to set
ceph config for target: global: failed to set keys
[mon_osd_backfillfull_ratio mon_osd_full_ratio mon_osd_nearfull_ratio]"

So we Specify the full ratios directly in the cephCluster CR under the
storage section. Use the specified value if given or else use the
defaults.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>