Merge specified placement with default placement instead of replacing it
Usually when specifying a placement someone specifies the part they
are interested in like a toleration or node affinity and do not specify
the rest. In this case we should merge the specified placement with the
default placement so that we use the default placement for the sections
that are not specified, instead of leaving them empty.
For node affinity, pod affinity and pod anti-affinity, if specified we
use the specified placement and default placement is not used.
For tolerations & topology spread constraints, we append the specified
placement to the default placement.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>