Switch to TSCs from pod anti-affinity for mds daemons
We have no reliable way to avoid scheduling 2 active MDS pods on
the same node. So instead of pod anti-affinity we can switch to TSCs
similar to how we use them for OSDs, with a max skew of 1. We can use
the rook_file_system: <fs-name> label selector so that all MDS daemon
(doesn't matter active or passive) of a file system are distributed
equally over the cluster.
Also TSCs are a better way to distribute pods across the cluster then
pod anti-affinity. Pod anti-affinity can't handle the case where more
than one pod need to be scheduled on the same node for example when
there are more than 1 active MDS. Tscs are easier to manage and
understand.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>