Always collect ocs must-gather irrespective of test result
After some discussion & thoughts, we decided to always collect the
ocs must-gather irrespective of test result. This would help a lot in
verifying the results of any fix whether or not that fix actually
results in a test failure. We acknowledge that this would result in
sight increase in the time taken to run the tests. But this would
help a lot in looking at the impact of any PR readily.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>