Return & requeue reconcile immediately after any client.Update() call
When client.Update() is called, the in-memory object is overwritten with
the API server response, losing any local changes to the status section.
This causes the status images section to be lost during the first
reconcile when the CR is updated due to either finalizer addition or
uninstall/cleanup policy annotations update. So if the same reconcile
would continue, it would result in a nil pointer dereference in the
noobaa & cephcluster ensureCreated functions when they try to access
the status images section.
To fix this we return with Requeue=true immediately after finalizer
addition or uninstall/cleanup annotations update. This prevents the nil
pointer dereference problem. For optimization, we do only one update
combination of finalizer addition & uninstall/cleanup annotations update
in a single call.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>