external: do not create metrics cephclient
in external mode we dont have admin privileges
and so we should avoid creating cephclient from the
ocs operator,
If needed in future we can create the cephclient
from the python script.

Currently external mode will rely on the fall back
mechanism to use the csi rbd secret for getting the
metrics exporter run.

Signed-off-by: parth-gr <partharora1010@gmail.com>