Get storageCluster as unstructured to avoid field default confusion
When the storageCluster is fetched and stored as a structured object,
the unspecified fields are set to their default values. Like bool fields
are set to false, string fields are set to empty string, etc.
Due to this we loose the ability to know the user's intent, whether
the bool field is set to false because it was not specified or because
it was explicitly set to false, whether the string field is set to empty
string because it was not specified or because it was explicitly set to
empty string.
If we get the storageCluster as unstructured, we can check if the field
exists in the unstructured object. If it does, it means the user has
explicitly set the field. If it does not, it means the field is not
specified by the user. This way we can know the user's intent and
avoid the confusion of default values.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>