Make possible to mark EnableCrushUpdates as false & deviceClass as empty
The enableCrushUpdates field is a bool field, and after we get the
storageCluster as a structured object it's impossible to know if the
field is set to false because it was not specified or because it was
explicitly set to false. Hence till now we were setting the field to
true always. Similarly for deviceClass field, on the structured sc it's
impossible to know if the field is set to empty string because it was
not specified or because it was explicitly set to empty string. Hence
till now we were setting the field to the default device class always.

Now we are getting the storageCluster as unstructured, and we can check
if the field exists in the unstructured object. If it does, it means the
user has explicitly set the field. If it does not, it means the field is
not specified by the user. This way the user can now make the field
enableCrushUpdates as false or deviceClass as empty string.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>