Pass some default block pool spec for .mgr, .nfs & replica-1 blockPools
Along with the default cephBlockPool we also always create the .mgr
blockPool. Also when nfs is enabled we also create the .nfs blockPool.
We have no way to specify the poolSpecs of these blockPools.
Hence for getting replica-2 working we need to pass the replicated size
spec if specified for the default cephBlockPool to the .mgr & .nfs
blockPools. Also to get multiple deviceClasses working we need to
pass the EnableCrushUpdates & deviceClass if specified for the default
blockPool to the .mgr &
.nfs blockPools.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>