Merge specified resource requirements with defaults acc to resource type
When users set resource requirements for a component, they usually only
provide values for the resource type they care about. For example, if
they run into memory issues with MDS, they might only set memory
requests/limits and won't specify anything for CPU.

Previously, our approach was to copy only the user-specified values.
This meant that if a resource type wasn't provided, it would be
dropped entirely. In the above example, the daemon would end up with no
CPU requests/limits at all.

To fix this, we now merge the user-specified resources with the
defaults. For each resource type that is not explicitly set by the user,
we apply the default values. For example, if only memory is specified,
default CPU values are applied. If only CPU is specified, default memory
values are applied.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>