Make possible to mark EnableCrushUpdates as false
The enableCrushUpdates field is a bool field, and after we get the
storageCluster as a structured object it's impossible to know if the
field is set to false because it was not specified or because it was
explicitly set to false. Hence till now we were setting the field to
true always.

Now we are getting the storageCluster as unstructured, and we can check
if the field exists in the unstructured object. If it does, it means the
user has explicitly set the field. If it does not, it means the field is
not specified by the user. This way the user can now make the field
enableCrushUpdates as false.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>