Set Maintenance Mode only if the cluster is configured for RDR
Maintenance Mode makes sense only on RDR clusters.
If we are not in a RDR env we should not bother with maintenance mode.
Previously we were setting maintenance mode based on the absence of
the rbd-mirror CR only, due to which maintenance mode was getting set
even in non rdr cluster.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>