fix block-only storagecluster deployment
storageclasses for rbd and cephfs are added by default to the
storageconsumer which ideally is not a problem since if we don't find
the spec we don't distribute it.

however, there is an unhandled panic when a nil interface is compared
with interface of a type of nil value and the checks pass where it
should fail.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>