Fix blockPool spec fields like failure domain not getting updated
The setDefaultDataPoolSpec is called on the actual poolSpec. In this the
fields like failure domain, device class, etc are already set.
So we end up not modifying them further even though they should now
have a new values based on the storageCluster. To solve this we now
start with an empty poolSpec in each reconcile. We preserve the existing
mirroring spec as it's handled by the mirroring controller.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>