Set Maintenance Mode only if RDR is configured & rbd-mirror CR is absent
Until now we were setting maintenance mode based on only the absence of
the rbd-mirror CR only, due to which maintenance mode was getting set
always even in non rdr cluster. To solve this we are now checking if
the cluster is configured for RDR & if the rbd-mirror CR is absent.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>