Roll back the CPU adjustment for IBM Z platform for now
The requirements from the IBM Z team was to use lower CPU limits & with
higher CPU requests. But this would result in the pods loosing their
k8s guaranteed QOS class. So we adjusted both the requests & limits for
IBM Z platform. But that resulted in loss of performance. So we are
rolling back the CPU adjustment for now. We need to revisit this when
we have a better understanding of QOS working with Limits & Requests.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>