Stop marking storageCluster Not Ready when operator is not upgradeable
The notion that StorageCluster is Not Ready to serve when the operator
is not upgradeable is incorrect. The operator can be not upgradeable for
reasons like mismatch of server & client versions etc which don't
necessarily affect storage operations.
Also mark the storageCluster as Error when the degraded condition is
true & mark it as Progressing when the progressing condition is true or
the available condition is false.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>