Make it possible to clear target size ratio on ocs created pools
Currently until 4.20 ocs operator applies a target size ratio of 0.49
on all the data pools it creates. Although possible to specify a diff
ratio but it's not possible for a user to not specify any ratio if
they want. We have seen that as the auto scaler has evolved having the
target size ratio is creating PG scaling issues and not having any
target size ratio actually works better. But as removing the target
size ratio completely is too big of a change we are giving an option to
make it possible to omit/remove the target size ratio from the pools if
a user wants. This will help support team quickly address PG scaling
related issues.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>